refactored to put find district form in a seperate component called FindDistrict
