+++ app/components/web-components/HartfordVotes/index.js
@@ -6,6 +6,7 @@ import HartfordCandidatesConversations from './HartfordCandidatesConversations'
 import HartfordLandingMenu from './HartfordLandingMenu'
 import { ModeProvider } from './phone-portrait-context'
 import { AddressProvider } from './user-address-context'
+import FindDistrict from './FindDistrict'
 
 const useStyles = createUseStyles({
   landingPage: {
@@ -49,6 +50,7 @@ const HartfordVotes = () => {
         <AddressProvider>
           <div className={classes.landingPage}>
             <HartfordLandingMenu key="hartford-menu" />
+            <FindDistrict />
             <HartfordLanding />
           </div>
         </AddressProvider>
